langgraph-react/pyproject.toml
[project]
name = "langgraph-react"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=2.1.4",
    "langchain-tavily>=0.1.6",
    "langchainhub>=0.1.21",
    "langgraph>=0.4.3",
    "langsmith>=0.3.42",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]
